Move PasswordCard to component/password-entry.js
