+++ src/component/password-entry.js
@@ -1,9 +1,66 @@
 import React, { Component } from 'react';
 import { View, TouchableOpacity, StyleSheet } from 'react-native';
 import PropTypes from 'prop-types';
-import Icon from './icon';
+import { FormSubText, FormStretcher } from './form';
+import { color, font } from './style';
 import { InputField } from './field';
-import { color } from './style';
+import Card from './card';
+import Icon from './icon';
+
+//
+// Password Card
+//
+
+const passwordStyles = StyleSheet.create({
+  card: {
+    maxHeight: 350,
+    maxWidth: 680,
+    paddingLeft: 45,
+    paddingRight: 45,
+    paddingBottom: 50,
+  },
+  newCopy: {
+    marginTop: 10,
+    maxWidth: 250,
+    color: color.blackText,
+    height: font.lineHeightSub * 2,
+  },
+});
+
+export const PasswordCard = ({
+  copy,
+  placeholder,
+  password,
+  onChangeText,
+  onSubmitEditing,
+  newCopy,
+  success,
+}) => (
+  <Card style={passwordStyles.card}>
+    <FormSubText>{copy}</FormSubText>
+    <FormStretcher>
+      <PasswordEntry
+        placeholder={placeholder}
+        value={password}
+        autoFocus={true}
+        onChangeText={onChangeText}
+        onSubmitEditing={onSubmitEditing}
+        success={success}
+      />
+      <FormSubText style={passwordStyles.newCopy}>{newCopy}</FormSubText>
+    </FormStretcher>
+  </Card>
+);
+
+PasswordCard.propTypes = {
+  copy: PropTypes.string,
+  placeholder: PropTypes.string,
+  password: PropTypes.string,
+  onChangeText: PropTypes.func.isRequired,
+  onSubmitEditing: PropTypes.func.isRequired,
+  newCopy: PropTypes.string,
+  success: PropTypes.bool,
+};
 
 //
 // Password Entry
